Fix integer type error #26

+++ LooperBuffer.h
@@ -216,9 +216,9 @@ public:
         writeHeads_[RIGHT_CHANNEL]->Stop();
     }
 
-    inline float ReadLeft(uint32_t position)
+    inline float ReadLeft(int32_t position)
     {
-        while (position >= kLooperChannelBufferLength)
+        while (position >= (int32_t) kLooperChannelBufferLength)
         {
             position -= kLooperChannelBufferLength;
         }
@@ -230,15 +230,15 @@ public:
         return buffer_[position];
     }
 
-    inline float ReadRight(uint32_t position)
+    inline float ReadRight(int32_t position)
     {
         position += kLooperChannelBufferLength;
 
-        while (position >= kLooperTotalBufferLength)
+        while (position >= (int32_t) kLooperTotalBufferLength)
         {
             position -= kLooperChannelBufferLength;
         }
-        while (position < kLooperChannelBufferLength)
+        while (position < (int32_t) kLooperChannelBufferLength)
         {
             position += kLooperChannelBufferLength;
         }
